-Added constructionabridged to construction.py.
