+++ app/models/samples/json/aperture_set.json
@@ -1,0 +1,1 @@
+{"type": "ApertureSetAbridged", "fixed_window_construction": "Exterior Window"}
