+++ app/models/samples/json/construction_door.json
@@ -1,0 +1,1 @@
+{"type": "OpaqueConstructionAbridged", "name": "Exterior Door", "layers": ["Metal Surface", "Insulation Board"]}
