+++ app/models/samples/json/construction_set.json
@@ -1,0 +1,1 @@
+{"type": "ConstructionSetAbridged", "name": "Construction Set", "wall_set": {"type": "WallSetAbridged", "exterior_construction": "Exterior Wall ASHRAE 2009"}, "floor_set": {"type": "FloorSetAbridged", "interior_construction": "Internal Floor"}, "roof_ceiling_set": {"type": "RoofCeilingSetAbridged", "exterior_construction": "Exterior Roof ASHRAE 2009"}, "aperture_set": {"type": "ApertureSetAbridged", "fixed_window_construction": "Exterior Window"}}
