+++ app/models/samples/json/construction_set_1.json
@@ -1,0 +1,1 @@
+{"type": "ConstructionSetAbridged", "name": "Construction Set 1", "wall_set": {"type": "WallSetAbridged", "exterior_construction": "Exterior Wall ASHRAE 2009"}, "floor_set": {"type": "FloorSetAbridged", "interior_construction": "Internal Floor"}}
