+++ app/models/samples/json/door_set.json
@@ -1,0 +1,1 @@
+{"type": "DoorSet", "exterior_construction": "Exterior Door"}
