+++ app/models/samples/json/floor_set.json
@@ -1,0 +1,1 @@
+{"type": "FloorSetAbridged", "interior_construction": "Internal Floor"}
