+++ app/models/samples/json/roof_ceiling_set.json
@@ -1,0 +1,1 @@
+{"type": "RoofCeilingSetAbridged", "exterior_construction": "Exterior Roof ASHRAE 2009"}
