+++ app/models/samples/json/wall_set.json
@@ -1,0 +1,1 @@
+{"type": "WallSetAbridged", "exterior_construction": "Exterior Wall ASHRAE 2009"}
